requirements.txt
numpy==1.19.2
pillow==9.4.0
scikit-image==0.19.3
tensorflow==2.4.0
tqdm==4.65.0
bm3d==4.0.1
tensorflow-addons==0.14.0